RDCC-2029 - moved all test cases to one file
