+++ src/functionalTest/java/uk/gov/hmcts/reform/cwrdapi/CaseWorkerRefFunctionalTest.java
@@ -1,12 +1,16 @@
 package uk.gov.hmcts.reform.cwrdapi;
 
 import com.google.common.collect.ImmutableList;
+import io.restassured.builder.MultiPartSpecBuilder;
+import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import io.restassured.specification.MultiPartSpecification;
 import lombok.extern.slf4j.Slf4j;
 import net.thucydides.core.annotations.WithTag;
 import net.thucydides.core.annotations.WithTags;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
+import org.apache.poi.util.IOUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.context.annotation.ComponentScan;
@@ -23,10 +27,15 @@ import uk.gov.hmcts.reform.cwrdapi.controllers.request.CaseWorkerWorkAreaRequest
 import uk.gov.hmcts.reform.cwrdapi.controllers.request.CaseWorkersProfileCreationRequest;
 import uk.gov.hmcts.reform.cwrdapi.controllers.request.UserRequest;
 import uk.gov.hmcts.reform.cwrdapi.controllers.response.CaseWorkerProfileCreationResponse;
+import uk.gov.hmcts.reform.cwrdapi.controllers.response.IdamRolesMappingResponse;
+import uk.gov.hmcts.reform.cwrdapi.util.CaseWorkerConstants;
 import uk.gov.hmcts.reform.cwrdapi.util.CustomSerenityRunner;
 import uk.gov.hmcts.reform.cwrdapi.util.FeatureConditionEvaluation;
 import uk.gov.hmcts.reform.cwrdapi.util.ToggleEnable;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -34,10 +43,14 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import static java.lang.String.format;
+import static org.apache.commons.collections.CollectionUtils.isEmpty;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.codehaus.groovy.runtime.InvokerHelper.asList;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.springframework.util.ResourceUtils.getFile;
 
 @ComponentScan("uk.gov.hmcts.reform.cwrdapi")
 @RunWith(CustomSerenityRunner.class)
@@ -46,11 +59,12 @@ import static org.junit.Assert.assertTrue;
 @Slf4j
 @TestPropertySource(properties = {"spring.config.location=classpath:application-functional.yml"})
 @SuppressWarnings("unchecked")
-public class CaseWorkerRefCreateFunctionalTest extends AuthorizationFunctionalTest {
+public class CaseWorkerRefFunctionalTest extends AuthorizationFunctionalTest {
 
     public static final String CREATE_CASEWORKER_PROFILE = "CaseWorkerRefController.createCaseWorkerProfiles";
     public static final String FETCH_BY_CASEWORKER_ID = "CaseWorkerRefController.fetchCaseworkersById";
     public static List<String> caseWorkerIds = new ArrayList<>();
+    public static final String CASEWORKER_FILE_UPLOAD = "CaseWorkerRefController.caseWorkerFileUpload";
 
     @Test
     @ToggleEnable(mapKey = CREATE_CASEWORKER_PROFILE, withFeature = true)
@@ -63,7 +77,7 @@ public class CaseWorkerRefCreateFunctionalTest extends AuthorizationFunctionalTe
 
         CaseWorkerProfileCreationResponse caseWorkerProfileCreationResponse =
                 response.getBody().as(CaseWorkerProfileCreationResponse.class);
-        caseWorkerIds = caseWorkerProfileCreationResponse.getCaseWorkerIds();
+        List<String> caseWorkerIds = caseWorkerProfileCreationResponse.getCaseWorkerIds();
         assertEquals(caseWorkersProfileCreationRequests.size(), caseWorkerIds.size());
     }
 
@@ -181,7 +195,7 @@ public class CaseWorkerRefCreateFunctionalTest extends AuthorizationFunctionalTe
     // this test verifies User profile are fetched from CWR
     @ToggleEnable(mapKey = FETCH_BY_CASEWORKER_ID, withFeature = true)
     public void shouldGetCaseWorkerDetails() {
-        if (caseWorkerIds.isEmpty()) {
+        if (isEmpty(caseWorkerIds)) {
             //Create 2 Caseworker Users
             List<CaseWorkersProfileCreationRequest> caseWorkersProfileCreationRequests = new ArrayList<>();
 
@@ -240,7 +254,7 @@ public class CaseWorkerRefCreateFunctionalTest extends AuthorizationFunctionalTe
     // this test verifies User profile are fetched from CWR when id matched what given in request rest should be ignored
     @ToggleEnable(mapKey = FETCH_BY_CASEWORKER_ID, withFeature = true)
     public void shouldGetOnlyFewCaseWorkerDetails() {
-        if (caseWorkerIds.isEmpty()) {
+        if (isEmpty(caseWorkerIds)) {
             List<CaseWorkersProfileCreationRequest> caseWorkersProfileCreationRequests =
                     caseWorkerApiClient.createCaseWorkerProfiles();
 
@@ -296,5 +310,130 @@ public class CaseWorkerRefCreateFunctionalTest extends AuthorizationFunctionalTe
                 .assertThat()
                 .statusCode(403);
 
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = true)
+    public void shouldUploadXlsxFileSuccessfully() throws IOException {
+        ExtractableResponse<Response> uploadCaseWorkerFileResponse =
+                uploadCaseWorkerFile("src/functionalTest/resources/CaseWorkerUserWithNoPassword.xlsx",
+                        201, CaseWorkerConstants.REQUEST_COMPLETED_SUCCESSFULLY,
+                        CaseWorkerConstants.TYPE_XLSX, ROLE_CWD_ADMIN);
+        CaseWorkerProfileCreationResponse caseWorkerProfileCreationResponse = uploadCaseWorkerFileResponse
+                .as(CaseWorkerProfileCreationResponse.class);
+        assertEquals(CaseWorkerConstants.REQUEST_COMPLETED_SUCCESSFULLY,
+                caseWorkerProfileCreationResponse.getCaseWorkerRegistrationResponse());
+        assertFalse(caseWorkerProfileCreationResponse.getCaseWorkerIds().isEmpty());
+        assertEquals(format(CaseWorkerConstants.RECORDS_UPLOADED,
+                caseWorkerProfileCreationResponse.getCaseWorkerIds().size()),
+                caseWorkerProfileCreationResponse.getMessageDetails());
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = true)
+    public void shouldUploadXlsFileSuccessfully() throws IOException {
+        ExtractableResponse<Response> uploadCaseWorkerFileResponse =
+                uploadCaseWorkerFile("src/functionalTest/resources/CaseWorkerUserXlsWithNoPassword.xls",
+                        201, CaseWorkerConstants.REQUEST_COMPLETED_SUCCESSFULLY, CaseWorkerConstants.TYPE_XLS,
+                        ROLE_CWD_ADMIN);
+
+        CaseWorkerProfileCreationResponse caseWorkerProfileCreationResponse = uploadCaseWorkerFileResponse
+                .as(CaseWorkerProfileCreationResponse.class);
+        assertEquals(CaseWorkerConstants.REQUEST_COMPLETED_SUCCESSFULLY,
+                caseWorkerProfileCreationResponse.getCaseWorkerRegistrationResponse());
+        assertFalse(caseWorkerProfileCreationResponse.getCaseWorkerIds().isEmpty());
+        assertEquals(format(CaseWorkerConstants.RECORDS_UPLOADED,
+                caseWorkerProfileCreationResponse.getCaseWorkerIds().size()),
+                caseWorkerProfileCreationResponse.getMessageDetails());
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = true)
+    public void shouldUploadServiceRoleMappingXlsxFileSuccessfully() throws IOException {
+        ExtractableResponse<Response> uploadCaseWorkerFileResponse =
+                uploadCaseWorkerFile("src/functionalTest/resources/ServiceRoleMapping_BBA9.xlsx",
+                        201, CaseWorkerConstants.IDAM_ROLE_MAPPINGS_SUCCESS, CaseWorkerConstants.TYPE_XLS,
+                        ROLE_CWD_ADMIN);
+
+        IdamRolesMappingResponse caseWorkerProfileCreationResponse = uploadCaseWorkerFileResponse
+                .as(IdamRolesMappingResponse.class);
+        assertTrue(caseWorkerProfileCreationResponse.getMessage()
+                .contains(CaseWorkerConstants.IDAM_ROLE_MAPPINGS_SUCCESS));
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = true)
+    public void shouldUploadServiceRoleMappingXlsFileSuccessfully() throws IOException {
+        ExtractableResponse<Response> uploadCaseWorkerFileResponse =
+                uploadCaseWorkerFile("src/functionalTest/resources/ServiceRoleMapping_BBA9.xls",
+                        201, CaseWorkerConstants.IDAM_ROLE_MAPPINGS_SUCCESS, CaseWorkerConstants.TYPE_XLS,
+                        ROLE_CWD_ADMIN);
+
+        IdamRolesMappingResponse caseWorkerProfileCreationResponse = uploadCaseWorkerFileResponse
+                .as(IdamRolesMappingResponse.class);
+        assertTrue(caseWorkerProfileCreationResponse.getMessage()
+                .contains(CaseWorkerConstants.IDAM_ROLE_MAPPINGS_SUCCESS));
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = true)
+    public void shouldReturn401WhenAuthenticationInvalid() {
+        Response response = caseWorkerApiClient.withUnauthenticatedRequest()
+                .post("/refdata/case-worker/upload-file/")
+                .andReturn();
+        response.then()
+                .assertThat()
+                .statusCode(401);
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = true)
+    public void shouldReturn403WhenRoleIsInvalid() throws IOException {
+        uploadCaseWorkerFile("src/functionalTest/resources/CaseWorkerUserWithPassword.xlsx",
+                403, null,
+                CaseWorkerConstants.TYPE_XLSX, "Invalid");
+    }
+
+    @Test
+    @ToggleEnable(mapKey = CASEWORKER_FILE_UPLOAD, withFeature = false)
+    public void shouldReturn403WhenUploadFileApiToggledOff() throws IOException {
+
+        String exceptionMessage = CustomSerenityRunner.getFeatureFlagName().concat(" ")
+                .concat(FeatureConditionEvaluation.FORBIDDEN_EXCEPTION_LD);
+
+        uploadCaseWorkerFile("src/functionalTest/resources/CaseWorkerUserWithNoPassword.xlsx",
+                403, exceptionMessage,
+                CaseWorkerConstants.TYPE_XLSX, ROLE_CWD_ADMIN);
+    }
+
+    private ExtractableResponse<Response> uploadCaseWorkerFile(String filePath,
+                                                               int statusCode,
+                                                               String messageBody,
+                                                               String header,
+                                                               String role) throws IOException {
+        MultiPartSpecification multiPartSpec =  getMultipartFile(filePath, header);
+
+        Response response = caseWorkerApiClient.getMultiPartWithAuthHeaders(role)
+                .multiPart(multiPartSpec)
+                .post("/refdata/case-worker/upload-file")
+                .andReturn();
+        response.then()
+                .assertThat()
+                .statusCode(statusCode);
+
+        return response.then()
+                .extract();
+    }
+
+
+    private MultiPartSpecification getMultipartFile(String filePath,
+                                                    String headerValue) throws IOException {
+        File file = getFile(filePath);
+        FileInputStream input = new FileInputStream(file);
+        MultiPartSpecBuilder multiPartSpecBuilder =  new MultiPartSpecBuilder(IOUtils.toByteArray(input))
+                .fileName(file.getName())
+                .header("Content-Type",
+                        headerValue);
+        return multiPartSpecBuilder.build();
     }
 }
